--- Front/SwifthubFront/src/app/components/feedback/add-feedback/add-feedback.component.html
@@ -1,30 +1,0 @@
-
-<form [formGroup]="addFeedbackForm">
-
-<div class="card mb-3 sticky-top" style="background-color: transparent; border: none;">
-    <div class="card-body">
-      <div class="row">
-        <div class="col">
-          <textarea class="form-control comment-textarea" rows="3" formControlName="description" name="feedbackDescription" placeholder="Write your feedback here..."></textarea>
-        </div>
-      </div>
-      <div class="row mt-2">
-        <div class="col-auto">
-          <select class="form-control" formControlName="rating" name="feedbackRating">
-            <option value="Great">Great</option>
-            <option value="Good">Good</option>
-            <option value="Neutral">Neutral</option>
-            <option value="Bad">Bad</option>
-          </select>
-        </div>
-        <div class="col-auto">
-          <button (click)="submitFeedback()" type="submit" class="btn btn-outline-light" style="color: #c8935e;">
-            Post Feedback
-          </button>
-        </div>
-      </div>
-    </div>
-  </div>
-</form>
-<p-toast></p-toast>
-
